requirements.txt
aniso8601==9.0.1
anyio==3.6.1
asgiref==3.5.2
astroid==2.11.6
attrs==19.3.0
autopep8==1.6.0
click==8.1.3
dill==0.3.5.1
Flask==2.1.2
flask-restx==0.5.1
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.8.0
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.6.0
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.6.1
platformdirs==2.5.2
pycodestyle==2.8.0
pyflakes==2.4.0
pylint==2.14.3
pyrsistent==0.18.1
pytz==2022.1
six==1.16.0
sniffio==1.2.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.0
typing-extensions==4.2.0
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.0

PyJWT~=2.4.0
passlib~=1.7.4
SQLAlchemy~=1.4.37